+++ src/pages/blog/[id]/index.tsx
@@ -4,7 +4,7 @@ import { Props as ArticleItemProps } from "@/ui/ArticleItem";
 import { Props as PageInfo } from "@/ui/ArticleTitle";
 import createImage from "@/utils/createImage";
 import { getBlocks, getDatabase, getPage } from "@/utils/notion";
-import { getSuggestArticles } from "./hooks";
+import { getArticles } from "@/utils/useGetArticles";
 
 export default function Article({
   id,
@@ -82,7 +82,24 @@ export const getStaticProps = async ({ params }: { params: { id: string } }) => 
     }),
   );
 
-  const suggestArticles = (await getSuggestArticles()) as ArticleItemProps[];
+  const getSuggestArticles = async () => {
+    const publicArticles = await getArticles();
+
+    const shuffleArray = (array: any[]) => {
+      for (let i = array.length - 1; i >= 0; i--) {
+        const tmp = Math.floor(Math.random() * (i + 1));
+        [array[i], array[tmp]] = [array[tmp], array[i]];
+      }
+      return array;
+    };
+
+    const suggestLength = 4;
+    const results = shuffleArray(publicArticles).slice(0, suggestLength);
+
+    return results as ArticleItemProps[];
+  };
+
+  const suggestArticles = await getSuggestArticles();
 
   return {
     props: {
